renaming bom to dependencies b/c it does the term bom a disservice
