+++ pkg/content/driver/k3s_test.go
@@ -13,7 +13,7 @@ func Test_newBom(t *testing.T) {
 	tests := []struct {
 		name    string
 		args    args
-		want    bom
+		want    dependencies
 		wantErr bool
 	}{
 		{
@@ -22,19 +22,19 @@ func Test_newBom(t *testing.T) {
 				kind:    "k3s",
 				version: "v1.22.2+k3s2",
 			},
-			want:    bom{},
+			want:    dependencies{},
 			wantErr: false,
 		},
 	}
 	for _, tt := range tests {
 		t.Run(tt.name, func(t *testing.T) {
-			got, err := newBom(tt.args.kind, tt.args.version)
+			got, err := newDependencies(tt.args.kind, tt.args.version)
 			if (err != nil) != tt.wantErr {
-				t.Errorf("newBom() error = %v, wantErr %v", err, tt.wantErr)
+				t.Errorf("newDependencies() error = %v, wantErr %v", err, tt.wantErr)
 				return
 			}
 			if !reflect.DeepEqual(got, tt.want) {
-				t.Errorf("newBom() got = %v, want %v", got, tt.want)
+				t.Errorf("newDependencies() got = %v, want %v", got, tt.want)
 			}
 		})
 	}
